entraid/scenario3/web-app/requirements.txt
Flask>=2.3.0
msal>=1.24.0
Flask-Session>=0.5.0
requests>=2.31.0
azure-identity>=1.15.0
